pyproject.toml
[project]
name = "crisis-sims"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.83.0",
    "google-genai>=1.64.0",
    "openai>=2.21.0",
    "pyyaml>=6.0.3",
]
